Feat: Course삭제 분기처리, lecture가 남아 있을 경우or 없을 경우

+++ src/hooks/mutation/useDeleteCourse.ts
@@ -8,16 +8,27 @@ interface IDeleteId {
   id: string;
 }
 
-interface IDeleteArr {
+const deleteCourses = async ({
+  deleteIdArray,
+  lectureCount,
+}: {
   deleteIdArray: IDeleteId[];
-}
+  lectureCount: number;
+}) => {
+  console.log(deleteIdArray);
 
-const deleteCourses = async (courseArr: IDeleteArr) => {
   try {
+    if (lectureCount + 1 !== deleteIdArray.length) {
+      return alert("아직 남아있는 강의가 있습니다");
+    }
     // 각 courseId에 대해 deleteDoc 작업을 수행하고, 결과 Promise들을 배열로 만듭니다.
-    const deletePromises = courseArr.deleteIdArray.map(({ id }) =>
-      deleteDoc(doc(db, "courses", id)),
-    );
+    const deletePromises = deleteIdArray.map(({ id, type }) => {
+      if (type === "course") {
+        deleteDoc(doc(db, "courses", id));
+      } else if (type === "lecture") {
+        deleteDoc(doc(db, "lectures", id));
+      }
+    });
 
     // Promise.all을 사용하여 모든 삭제 작업이 완료될 때까지 기다립니다.
     await Promise.all(deletePromises);
